Backend/requirements.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
python-decouple
requests
storage3
supabase
